Rename the "Vat"
