+++ src/mod.ts
@@ -7,7 +7,7 @@ export * from './faucets/mod.js';
 export * from './infusions/mod.js';
 export * from './joints/mod.js';
 export * from './mix/mod.js';
-export * from './refaucets/mod.js';
+export * from './vats/mod.js';
 export * from './seeps/mod.js';
 export * from './sink-all.js';
 export * from './sink.js';
