--- src/refaucets/mod.ts
@@ -1,3 +1,0 @@
-export * from './create-refaucet.js';
-export * from './refaucet.js';
-export * from './with-inflow.js';
